refactor java implementation, general cleanup
